requirements.txt
torch
torchvision
transformers
pillow
gradio
plotly
numpy